+++ arch/src/instructions/comparison.rs
@@ -82,7 +82,7 @@ pub fn gte(mode: Addressing, operand: &[Operand; 2], cpu: &mut CPU) -> Instructi
 
 pub fn srz(mode: Addressing, operand: &mut [Operand; 2], cpu: &mut CPU) -> InstructionResult {
     if let Addressing::Direct = mode {
-        operand[0].write_word(cpu, u8::from(cpu.status_register.zero))?;
+        operand[0].write_word(cpu, u32::from(cpu.status_register.zero))?;
         Ok(())
     } else {
         Err(Interrupt::IllegalInstruction)
@@ -91,7 +91,7 @@ pub fn srz(mode: Addressing, operand: &mut [Operand; 2], cpu: &mut CPU) -> Instr
 
 pub fn src(mode: Addressing, operand: &mut [Operand; 2], cpu: &mut CPU) -> InstructionResult {
     if let Addressing::Direct = mode {
-        operand[0].write_word(cpu, u8::from(cpu.status_register.carry))?;
+        operand[0].write_word(cpu, u32::from(cpu.status_register.carry))?;
         Ok(())
     } else {
         Err(Interrupt::IllegalInstruction)
@@ -100,7 +100,7 @@ pub fn src(mode: Addressing, operand: &mut [Operand; 2], cpu: &mut CPU) -> Instr
 
 pub fn sro(mode: Addressing, operand: &mut [Operand; 2], cpu: &mut CPU) -> InstructionResult {
     if let Addressing::Direct = mode {
-        operand[0].write_word(cpu, u8::from(cpu.status_register.overflow))?;
+        operand[0].write_word(cpu, u32::from(cpu.status_register.overflow))?;
         Ok(())
     } else {
         Err(Interrupt::IllegalInstruction)
